test(match): 매치 관련 서비스와 도메인 테스트 코드 작성 및 버그 수정

+++ src/main/java/com/shootdoori/match/dto/MatchCreateRequestDto.java
@@ -9,7 +9,7 @@ public record MatchCreateRequestDto(
     LocalDate preferredDate,
     LocalTime preferredTimeStart,
     LocalTime preferredTimeEnd,
-    Integer preferredVenueId,
+    Long preferredVenueId,
     SkillLevel skillLevelMin,
     SkillLevel skillLevelMax,
     Boolean universityOnly,
